requirements.txt
transformers
torch
python-dotenv
accelerate
bitsandbytes
peft
pymilvus
